implement delete city, item_city ui refactoring

+++ app/src/main/java/com/oliver/weatherapp/data/local/dao/CitiesDao.java
@@ -20,6 +20,6 @@ public interface CitiesDao {
     void insert(CityEntry city);
 
     @Query("DELETE FROM cities WHERE id = :id")
-    void deleteCity(int id);
+    void deleteCity(long id);
 
 }
